requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
minio
pymilvus
milvus-cli
redis
requests
dramatiq[redis]
python-dotenv
python-docx
pydantic-settings
pydantic[email]
passlib[bcrypt]
python-jose
python-multipart
puremagic
cryptography
openai
beautifulsoup4
lxml
PyMuPDF
olefile
jinja2
transitions
zipstream
psutil